Enable "-Wunique-object-duplication"
We've fixed all the problems in first-party code, and used the warning
suppression file to disable them in third-party code. Thus we're able
to finally turn this warning on.

We only enable it for component builds since it's only applicable there,
and release builds give false positives due to building with
-fvisibility=hidden.

Bug: 368047728
Change-Id: I8a44618127ec0f16efa38e7a0a8edbb2aa0a5ec0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6604255
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1468978}