Invalidation tracing: Move diagnostic event to debug category
The UseStyleRuleMapForSelectorStats trace event is not needed for style
invalidation analysis. Rather, it's a diagnostic event which was added
to measure how well the rule-to-stylesheet mapping is working. Due to
how frequently we perform lookups in that mapping, we were generating a
large number of these events, in some cases enough to exhaust DevTools'
capacity for saving a trace. This CL addresses that issue by moving the
event to a separate debug category.

Bug: 417227406, 337076014
Change-Id: I947fb41526ac05d4b267a45eb6422e51a000472c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6582305
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469003}