[gpu] Remove DumpWithoutCrashing for D3D11 NV12 texture creation
We added DumpWithoutCrashing to investigate cases where client gets
D3D11 backing factory but fails texture creation for NV12 textures
due to sizes that are too large. We found that media's
GMBVideoFramePool [1] is the client where the shared image is created
for video frame and has a large size.

Ideally we should query for a max texture size that is supported by
d3d11 for NV12 format and store it in shared image caps but this is
very platform dependent and not so straight-forward. This change
remove the added crash keys and dumps.

[1] https://source.chromium.org/chromium/chromium/src/+/main:media/video/gpu_memory_buffer_video_frame_pool.cc;drc=6099daedae9edf5b9ccbfb8267e108aaaf87371b;l=1243

Bug: 374461672
Change-Id: I6e56022a6272c163d57fecc205884bfa5a7969ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6617144
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469019}