Simplify inclusive language dirs script
This patch simply simplifies the inclusive language presubmit
script to remove the counts, which are no longer consumed by the
depot tools presubmit script.

This lowers the amount of noise in the inclusive language presubmit
exempt dirs file and simplifies the script that generates this file.

Change-Id: I8c27706984667634bb99bcca716fb27b1be64df7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6615538
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469068}