[AnchoredPopupWindow] Allow ignoring minimal touchable size limit
AnchoredPopupWindow has an undocumented behavior that it refuses to show
a popup smaller than 50dp/50dp just because it is too small to tap.

As we use AnchoredPopupWindow to implement an extension popup on Desktop
Android, we have to support popups as small as 25px/25px since it is
formerly documented in the public extension API reference:
https://developer.chrome.com/docs/extensions/reference/api/action#popup

In fact, the "minimal touchable" restriction does not make sense on
Desktop Android where users typically use a mouse or a track pad.

This patch adds an option to AnchoredPopupWindow to allow popups smaller
than the threshold. It is disabled by default, but callers that want to
avoid the restriction can turn it on by calling setAllowNonTouchableSize.

Bug: 385987224
Change-Id: If8e241daae5fde0d833c8c5765b95e402ea6b22a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6615379
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469126}