Move origin matcher to its own component
There is an API in content to filter origins by a list of origin
patterns using scheme_host_port_matcher. We had to pass those types over
IPC. The origin_matcher already does everything we needed but we decided
to first hand roll a solution to separate the work out because
origin_matcher is sitting in a component that content can not directly
rely on.

We initially moved this origin_matcher type to content but we got
advised by the content owners to move this type to its own component
where it can be shared between the component and content.

Bug: 407420300
Change-Id: I67437dcba082f6efca3522613c1cc8a60ff1e597
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6616500
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Commit-Queue: Rupert Wiser <bewise@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469138}