[iOS]Replace SigninPresenter by SigninPromoViewMediatorDelegate
This will allow, in the next CL, to add a method in this protocol only.

This CL also edit a test. It was testing that a weak variable was
automatically set to nil. There is no guarantee by obj-c, this is not
something that we should have tested. Also, otherwise, moving the
property to a weak instance variable was breaking the test.

Bug: none
Change-Id: I0f895383b6e945de82c28249f691683719292f03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6611462
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469212}