[iOS][safari-import] Show Safari import on FRE and reminder
This CL takes care of displaying the Safari import workflow on FRE and
on user reminder. In case it should be displayed on FRE, it waits for
the guided tour (if enabled) to finish to begin presenting.

Also refactored the guided tour and the Safari import into a separate
screen provider, which is the newly added `FirstRunPostActionProvider`.

Fixed: 420703280
Change-Id: I57c5a90c1f563baf13f8f9ace5f854c4598af398
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6619107
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Aliona Dangla <alionadangla@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469216}