Capture the default output device in CatapAudioInputStream
This CL changes the behavior of CatapAudioInputStream so that the
output of the default output device is captured. This is true for
all device IDs except `kLoopbackAllDevicesId` for which we will
capture the system audio of all output devices.

The old behavior of always capturing all system audio regardless
of the output device can be restored by enabling the feature
MacCatapCaptureAllDevices.

Bug: 415953612
Change-Id: Ic2c8fb4c4d7179d6b76a90a2c6113d42c9b40129
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6620078
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Johannes Kron <kron@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469253}