Remove tasks from ActorCoordinator.
This CL is a refactor with no intended behavior change. It removes the
concepts of tasks from ActorCoordinator, which no longer do anything.
This removes a bunch of boilerplate.

There is currently a special case where a new tab needs to be made as
the first action for some tasks. This is now done inline, synchronously,
in AiDataKeyedService and GlicActorController.

Note that this CL is intentionally mirroring structure between
GlicActorController and AiDataKeyedService as the former will be
replaced by the latter.

Change-Id: I8b477ffcd2b6f2ca2739e23cde52079496640cd4
Bug: 411462297
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6605934
Reviewed-by: Ryan Sturm <ryansturm@chromium.org>
Commit-Queue: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469333}