Fix crash due to BrowserBoundKey not present.
When StrongBox is not available, before this change
BrowserBoundKeyStoreAndroid created a BrowserBoundKeyAndroid with a null
JavaRef, and subsequently a crash occurs during the
GetPublicKeyAsCoseKey() call. After this change
BrowserBoundKeyStoreAndroid returns nullptr (instead of creating a
BrowserBoundKeyAndroid with a null JavaRef), thus avoiding the crash.

Browser bound keys are behind the flag
chrome://flags/#enable-secure-payment-confirmation-browser-bound-key.

Bug: 377278827, 421933357
Change-Id: Ib378f79f0f2afb57a79c6de5a104eab77f6b4bf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6612120
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Commit-Queue: Slobodan Pejic <slobodan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469341}