[InputVizard] Do not try creating child surface if parent is null
We had received some crashes from field where the clients are crashing
with `nullptr passed as parent argument`, which seems to be happening
when parent can be nullptr. This is coming only from handful of clients.
In this change we are just early returning from the method and log in
uma if parent surface was null for some reason.

Bug: 40073296
Change-Id: Ia62c731565d43930505b5da74218d6fa39fdcffb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6616621
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Commit-Queue: Kartar Singh <kartarsingh@google.com>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469361}