blockfile: trying to deal with squished hash tables, take two.
My previous attempt at dealing with the caused-by-buggy-canary
corruption by just not DCHECK'ing on it was wrong; since while it's OK
to ignore items being in wrong buckets when opening, it gets weird when
it's time to evict them and they're not linked in the right place.

So this switches to an approach mmenke@ suggested: if we detect a
problem, blow the cache away. For simplicity, this is done at the next
open (since doing it while open, while theoretically supported, is quite
a lot more dangerous). To monitor this, added a new enum to
DiskCache.0.Error to show when this discard at open happens, and also
add a metric for hash mismatch in particular.

Bug: 421211228, 40063884
Change-Id: If9512ccc818aa41840bcfac426497c4ca2ee701b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6617005
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469392}