[Extensions] Pass ProcessManager to ServiceWorkerState constructor
This change refactors ServiceWorkerState to store a pointer to the
ProcessManager instead of passing it as an argument to multiple methods.

This is a necessary step for further refactoring of
ServiceWorkerTaskQueue. Eventually, ServiceWorkerState will be entirely
responsible for handling DidStartWorkerForScope and other related
signals. So we can't rely on SWTQ passing the ProcessManager on demand.

Design doc:
https://docs.google.com/document/d/1_4xu81Mffh2GCeovn_R9auYl-qfIkBnHt_MEzTMPxzc/edit?usp=sharing

Bug: 40276609
Change-Id: Ice037615bc11de97c37d646868aade3817090afd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618427
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Andrea Orru <andreaorru@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469395}