[accessibility] Don't remove unknown trees from the registry
AXActionHandlerBase sets itself as the handler for a given tree ID. At
construction, an instance may not yet know the tree ID for which it is
the handler -- the tree ID will be "unknown". Since it never registers
itself as the handler for an unknown tree, there is no need for it to
remove itself from the registry when its tree ID was unknown.

Bug: none
AX-Relnotes: n/a.
Change-Id: I843ccdf3a7a0b653bcf8b2473fa4125134696320
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618651
Auto-Submit: Greg Thompson <grt@chromium.org>
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469403}