Correctly handle no-target case in ScrollTool
After https://crrev.com/c/6579529, the scroll tool wasn't correctly
expecting a target to be empty. This wasn't caught by tests because they
also incorrectly initialized the proto with a DocumentIdentifier which
implicitly creates a target. Additionally, this doesn't trip the CHECK
that was looking for a nullptr target because the proto bindings return
a default ActionTarget rather than a nullptr in this case.

Fixed: 422243157
Change-Id: Ief50ceebb54964954be3065c0f6685408aa4100d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618660
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469414}