[WV] Catch RuntimeException from cache APIs
As part of the project to increase WebView's HTTP cache size we are
calling two framework APIs related to cache usage and quota.
Unfortunately, these APIs require the STORAGE_INTERNAL permission which
the Rubidium runtime does not have and throws a SecurityException
causing a crash. This caused a bunch of failures on the Rubidium
builders, see crbug.com/422174715. This CL makes the change to catch
RuntimeException (parent class of SecurityException) from these API
calls.

We also update the field trial config to test that we do not see the
same regressions on the builders.

Bug: 417434323
Change-Id: Ia83fefd7124310eaba2884b62a5759a9c69d5734
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6620524
Commit-Queue: Abhijith Nair <abhijithnair@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469430}