Fix read aloud not showing up beyond the first load.
The was broken in this CL:
crrev.com/3e37737aa5dd9988c545b44c731e5d6e21fa24a3

Prior to this CL, mHasReadAloudInserted was set to true/false
based on the initial visibility of what was updated on the inflated
Menu object. So in this previous state, it only relied on dynamically
creating the PropertyModel sometimes and other times it would use
the MenuItem -> PropertyModel conversion logic. The above CL did
not reset the inserted value, so after it was set to true once it
would never get added back.

Bug: 422424108
Change-Id: I3ed5f82a09002725da9ab4bdac75ea472bd5fd3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6621686
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469438}