[FedCM] Add check before running on_account_selection_
This is a tentative fix where we are crashing due to
`CHECK(on_account_selection_)` on Android. I suspect this is being
caused by OnDismiss having been invoked before OnAccountSelected. So
instead of directly checking for the presence of
`on_account_selection_`, we check for the presence of `on_dismiss_` to
confirm the hypothesis.

Bug: 398778802
Change-Id: I58ebc08164068745981c0dd3d9f1b6a2575de0cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6621582
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469444}