Local Storage: Replace CHECK in d'tor with DWOC
In crrev.com/c/6418502 we added a CHECK in ~LocalStorageImpl. This was
intended to convert a UaF in ~StorageAreaHolder into an ASSERT. This
would fix the UaF while preserving crashes. In fact, this has led to an
increase in crash frequency that is deemed Stable blocking. So, this
change uses DumpWithoutCrashing to continue collecting crash dumps while
avoiding the additional crashes caused by the addition of the CHECK.

Bug: 396030877
Change-Id: Ic8230ca0acb31dc60e8b59ad02ba90e965af027a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618271
Reviewed-by: Evan Stade <evanstade@microsoft.com>
Commit-Queue: Rahul Singh <rahsin@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1469473}