Populate Trust Anchor IDs from intersection of DNS and trusted TAIs
This CL glues together prior CLs to intersect the server-advertised
Trust Anchor IDs from DNS with the TAIs trusted by
TrustStoreChrome. SSLConnectJob uses this intersection to populate the
respective field on SSLConfig, when a feature flag is enabled.

The code to compute the intersection and translate it into wire format
is a static helper method on CertVerifier, so that it can be usable
from QuicChromiumClientSession in a subsequent CL (as well as from
SSLConnectJob in this CL).

Bug: 415720499
Change-Id: Icca4ad8ae9d09c9d8fcd19d483ee5e8b8bcbc48e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6578737
Reviewed-by: Nick Harper <nharper@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469478}