mac: fix event routing in titled window
In titled windows, the bottom part of the window stops receiving mouse
event. This was because the code misuses window or superview coordinate
as the view local coordinate for hit testing. This mistake exists in
both BridgedContentView (which hosts a toolkit-views tree) and
RenderWidgetHostViewCocoa (which hosts a content area).

Fixed: 422421135
Change-Id: I82fe84aede4725c2a428aacbfb3ea1de0ac06b5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6621334
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Keren Zhu <kerenzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469485}