[iOS] Refactor the First Run directory to expose public APIs
This CL refactors the BestFeaturesItem class to a public API. The helper
functions IsWelcomeBackEnabled and MarkWelcomeBackFeatureUsed are also
exposed. This refactor is to prevent code in model/ directories from
depending on code in the first_run/ui_bundled directory, which is
required to mark features as used for the Welcome Back promo.

Bug: 410072328
Change-Id: I6675aafffb2849b0324816b0fb57a65b1e4a1ef9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618426
Reviewed-by: Hira Mahmood <hiramahmood@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Briana McClure <bmcclure@google.com>
Cr-Commit-Position: refs/heads/main@{#1469500}