[NativeFrameView] Consolidate test classes into one configurable class
Consolidates HitTestNonClientFrameView and MinimumSizeFrameView which
were defined in tests into a common configurable test class.

The high level goal with this CL is to reduce the number of subclasses
of NativeFrameView which in turn reduces the scope of changes for a
larger refactor at a later time.

Change-Id: Ia8f7cd3bd526ad59e5238f71883e35721830a5f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6613673
Reviewed-by: Robert Liao <robliao@chromium.org>
Commit-Queue: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469518}