Make screenshot_sync dpr-aware
Makes the GPU screenshot_sync dpr-aware when calculating the boundaries
that should be checked. Previously, this implicitly assumed a DPR of 1.

This is specifically being done to fix the tests on corsola devices
running Android Desktop, which have a DPR of <1, which resulted in
background pixels being sampled. However, this fix is also beneficial
for devices with DPRs >1, as the entire image should be properly sampled
now.

Bug: 413427416
Change-Id: I9e82750e26fcd13301e01916384dbf2096367807
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6619594
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469536}