Adjust share button colors.
These two intractable buttons (share/manage tab group, and send
feedback) were both using colorPrimaryContainer. While this worked for
some more common color schemes, on some OEMs, in monochrome mode, and/or
with different contrast ratio settings we did not get good contrast with
the foreground text and/or the background scrim. This was partly because
we should only have been using colorOnPrimaryContainer with
colorPrimaryContainer, per material guidance. And it was partly because
colorPrimaryContainer has more variation than we thought.

To fix this, we have switched the tab group manage button to use
colorOnPrimaryContainer (text) since it doesn't need to contrast with a
scrim. The send feedback button was changed to use colorPrimary (bg) and
colorOnPrimary (text) as it does need to contrast with a scrim.

Bug: 411459427
Change-Id: I02a6207f06b10159ece97877b23e8f08a3a46df0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618248
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Hailey Wang <haileywang@google.com>
Cr-Commit-Position: refs/heads/main@{#1469547}