Add gsl::Owner annotation to base::expected
This attribute tells clang's dangling pointer analysis that the type
serves as an "owner" for pointer-like types such a string_view which
appear in its template arguments. Having the attribute causes clang
to warn on conversions that are unsafe due to lifetime issues.

Due to a bug/limitation of the analysis, the warning currently only
detects problem if the pointer-like type is the first template argument.
That is, it will fire for conversions to `expected<string_view, T>` but
not to `expected<T, string_view>`.

Bug: 421941585
Change-Id: Ia87adb71a3e02f930c0d32f991dcb537ed541003
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6618855
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1469561}