Spanify net/filters.
This removes all remaining allow_unsafe_* pragmas in net/filters, which
are only in test files.

Since that includes updating MockSourceStream's public API, this
also updates tests of FilterSourceStreams that reside outside net,
and spanifies them as well.

In addition to addressing function calls that need the pragma, it
additionally removes most calls to std::string::append(char*, size_t),
which don't trigger the warning, but are morally unsafe. It does leave
in a call or two to that method when using const C arrays.

Bug: 40284755
Change-Id: I9cbf000e29afc8011321f4bafa7a24e1cf76bb0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6611280
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469569}