Add logging in case of failure in ActorCoordinator
There are cases (such as the linked bug) where a failure occurs without
invoking the tool in ToolCoordinator (which logs the result). Log
failures in ActorCoordinator as well - this may be redundant in cases
where the ToolController is invoked but will help debugging.

Also update VLOG to ACTOR_LOG for consistency (no-op).

Bug: 422243157
Change-Id: Ia87af1e9d562c4bc482376cb9a2a7c216e9d0121
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6621186
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1469613}